Added volume command (without command id); Updated price command to include percentage change; Refactored utility functions for commands;
