+++ .gitignore
@@ -4,3 +4,5 @@ transpiled
 worker
 .vscode
 .idea
+._*
+**/.DS_Store
